Add support for quilted packaging
In this model, we don't actually have most of the packaged files in
git.  Instead, we have only a bare `debian` directory.  The
debian/archive.sh file downloads and packages the orig tarball in the
parent directory.  The debian/bootstrap.sh file extracts the files
into the current directory and applies the quilt patches.